Add files via upload
This Dashboard is a visual of the different features for visual creation and this includes the usage of various data sets of a companies 'Laptop sales and Production Products in line with stakeholder concepts and insights. Also, the background of the visuals was with numeric techniques such as 'Normal or Quick Measure tools'.