Add files via upload
This report was designed for a financial project that entails visual cards and charts and this includes total COGS for each segment, the total sales for each product, the relationship between Sales and COGS and the proportion of the total sales for each product. furthermore, all the charts  display different visuals of asset information.